requirements.txt
# FastAPI Core
# Note: If fastapi>=0.100.0 is already installed, it may require pydantic>=2.0
# Adjust these constraints based on your installed FastAPI version
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
# Using flexible constraints - will work with pydantic 1.x or 2.x depending on FastAPI version
pydantic>=1.10.0
pydantic-settings>=1.10.0

# Database
sqlalchemy>=1.4.0,<2.0.0
alembic>=1.10.0,<2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# AI/ML Core Libraries (Python 3.8 compatible)
numpy>=1.19.0,<1.25.0
pandas>=1.3.0,<2.0.0
scikit-learn>=1.0.0,<1.4.0
scipy>=1.7.0,<1.11.0

# Deep Learning (Python 3.8 compatible)
# Note: tensorflow 2.13.1 doesn't exist for macOS ARM64, so we cap at 2.13.0
tensorflow>=2.8.0,<=2.13.0
torch>=1.12.0,<2.0.0
transformers>=4.20.0,<4.38.0
sentence-transformers>=2.2.0,<2.3.0

# Computer Vision
opencv-python>=4.5.0,<4.10.0
Pillow>=9.0.0,<11.0.0
imageio>=2.20.0,<2.34.0

# NLP
spacy>=3.4.0,<3.8.0
nltk>=3.7.0,<3.9.0
textblob>=0.17.0,<0.18.0

# Time Series & Forecasting
statsmodels>=0.13.0,<0.15.0
prophet>=1.1.0,<1.2.0
pmdarima>=2.0.0,<2.1.0

# Recommendation Systems
implicit>=0.5.0,<0.7.0
scikit-surprise>=1.1.0,<1.2.0

# Feature Engineering
category-encoders>=2.5.0,<2.7.0
feature-engine>=1.5.0,<1.7.0

# Model Training & Evaluation
xgboost>=1.6.0,<2.0.0
lightgbm>=3.3.0,<4.0.0
catboost>=1.1.0,<1.3.0
imbalanced-learn>=0.10.0,<0.11.0

# Vector Databases & Embeddings
faiss-cpu>=1.7.0,<1.8.0
chromadb>=0.3.0,<0.5.0

# API & HTTP
httpx>=0.24.0,<0.27.0
aiohttp>=3.8.0,<4.0.0
requests>=2.28.0,<3.0.0

# Utilities
python-dotenv>=0.19.0,<1.1.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.0,<1.8.0
python-dateutil>=2.8.0,<3.0.0
pytz>=2022.0

# Monitoring & Logging
prometheus-client>=0.16.0,<0.20.0
structlog>=22.0.0,<25.0.0

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.20.0,<0.24.0

# CORS (Note: python-cors doesn't exist, using fastapi's built-in CORS)
# CORS is handled by fastapi.middleware.cors.CORSMiddleware

